+++ src/pages/Detail/components/comment/CommentItem.tsx
@@ -1,7 +1,7 @@
 import { Button } from "@/components/ui/button";
 import { generateDateString } from "@/lib/generateDateString";
 import { CommentItem as CommentItemResponse } from "@/pages/Detail/apis/useFetchCommentList";
-import CommentDeletionButton from "@/pages/Detail/components/comment/CommentDeletionButtons";
+import CommentDeletionButton from "@/pages/Detail/components/comment/CommentDeletionButton";
 import CommentUpdateForm from "@/pages/Detail/components/comment/CommentUpdateForm";
 import { useState } from "react";
